+++ kube-client/src/api/util.rs
@@ -1,6 +1,6 @@
 use crate::{
     api::{Api, Resource},
-    Result,
+    Error, Result,
 };
 use kube_core::util::Restart;
 use serde::de::DeserializeOwned;
@@ -11,7 +11,7 @@ where
 {
     /// Trigger a restart of a Resource.
     pub async fn restart(&self, name: &str) -> Result<K> {
-        let mut req = self.request.restart(name)?;
+        let mut req = self.request.restart(name).map_err(Error::BuildRequest)?;
         req.extensions_mut().insert("restart");
         self.client.request::<K>(req).await
     }