+++ kube-client/src/client/tls.rs
@@ -40,10 +40,13 @@ pub mod native_tls {
     // TODO Replace this with pure Rust implementation to avoid depending on openssl on macOS and Win
     fn pkcs12_from_pem(pem: &[u8], password: &str) -> Result<Vec<u8>> {
         use openssl::{pkcs12::Pkcs12, pkey::PKey, x509::X509};
-        let x509 = X509::from_pem(pem)?;
-        let pkey = PKey::private_key_from_pem(pem)?;
-        let p12 = Pkcs12::builder().build(password, "kubeconfig", &pkey, &x509)?;
-        let der = p12.to_der()?;
+        // TODO These are all treated as the same error. Add specific errors.
+        let x509 = X509::from_pem(pem).map_err(Error::OpensslError)?;
+        let pkey = PKey::private_key_from_pem(pem).map_err(Error::OpensslError)?;
+        let p12 = Pkcs12::builder()
+            .build(password, "kubeconfig", &pkey, &x509)
+            .map_err(Error::OpensslError)?;
+        let der = p12.to_der().map_err(Error::OpensslError)?;
         Ok(der)
     }
 }